pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "helper-agent"
version = "0.1.0"
description = "LangGraph Helper Agent for documentation Q&A"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "chromadb==1.3.5",
    "google-genai==1.52.0",
    "langgraph==1.0.4",
    "langchain==1.1.0",
    "langchain-google-genai==3.2.0",
    "langchain-text-splitters==1.0.0",
    "pandas==2.3.3",
    "tqdm==4.67.1",
    "tenacity==9.1.2",
]

[project.optional-dependencies]
dev = [
    "black>=24.0",
    "isort>=5.13",
    "flake8>=7.0",
    "mypy>=1.8",
    "pytest>=8.0",
]

[tool.setuptools]
packages = ["helper_agent"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
include_trailing_comma = true
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
